A little better makeiso.sh

+++ makeiso.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
+
 make kernel
+if [ $? -ne 0 ]
+then
+  exit
+fi
+
 cp bin/yuunos.bin iso/boot
 rm iso/yuunos.iso
 grub-mkrescue -o iso/yuunos.iso iso
